Sat, Dec 11, 2021, 4:20 AM +00:00

+++ Project/app.py
@@ -285,6 +285,9 @@ def upload():
             flash('Photo name already in use in this gallery')
             return redirect("/edit?g=" + gallery_id)
         f = request.files['photo']
+        if f.filename == '':
+            flash('Please provide ".jpeg" file to upload')
+            return redirect("/edit?g="+gallery_id)
         f.save(secure_filename(f.filename))
         insert_picture(f.filename.replace(" ", "_"), photo_name, gallery_id, session["user_id"])
         os.remove(f.filename)
